Move tasklist to tasks package
