+++ src/main/java/lex/parser/command/TodoCommand.java
@@ -1,6 +1,6 @@
 package lex.parser.command;
 
-import lex.TaskList;
+import lex.tasks.TaskList;
 import lex.tasks.Todo;
 import lex.ui.Ui;
