+++ src/main/java/lex/storage/Storage.java
@@ -8,7 +8,7 @@ import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 
-import lex.TaskList;
+import lex.tasks.TaskList;
 
 public class Storage {
     private final String filePath;
